Implement savePick function
Change binding method of handlePickChange method from binding in
constructor to binding in jsx with (this, args) in order to add the
selected team name as an argument to handlePickChange, which then
invokes savePick on that team name. The savePick method uses spread
operator to update the array to include the team name selected.